feat(metal): implement ComputeDevice; fix Apple build; drop out-of-scope blas
Bring Metal onto the unified compute surface -- the fourth backend to
implement core::ComputeDevice -- and fix the crate so it compiles for Apple
at all.

Metal compute:
- New compute::Context: compute pipeline from a compiled .metallib,
  shared-storage buffers, and dispatch. Implements ComputeDevice
  (device_buffer / upload / download / pipeline / dispatch / buffer_len).
- Apple unified memory means upload/download are a memcpy through the
  buffer's contents() pointer -- no staging.
- Metal sets threadgroup size at dispatch (the other three backends declare
  it in the shader), so the trait's dispatch assumes a 1-D group of 64;
  Context::dispatch_sized takes an explicit size for other geometries.
- Verified by cross-checking to aarch64-apple-darwin (default and mps
  features). Not run -- this workspace has no macOS host.

Fix Apple build (pre-existing, never compiled):
- blas.rs was compiled unconditionally on Apple but used metal's MPS types
  without enabling the mps feature, and targeted an mps::Matrix* API that no
  longer exists in metal 0.32. It is also an MPS-backed GEMM -- a
  workload-level kernel the crate's low-level-only scope excludes -- so it
  is removed rather than repaired; GEMM belongs above the driver line.
- Drop redundant inner cfg attributes and an unnecessary cast flagged by
  clippy on the Apple target.

Docs: core trait, umbrella, and README now list Metal as the fourth
implementer with its metallib/threadgroup notes; support matrix updated.

Verification: native workspace build, all-features tests, clippy
--all-targets, core doctest, and the Apple cross-check are all clean.

+++ crates/ironaccelerator-metal/src/drv.rs
@@ -1,8 +1,6 @@
 //! Safe Metal driver layer. Apple-only — the module is empty on other hosts
 //! so the non-Apple scaffold keeps building.
 
-#![cfg(target_vendor = "apple")]
-
 use ironaccelerator_core::{Error, Result};
 use metal::{Buffer as MtlBuffer, CommandQueue, Device as MtlDevice, MTLResourceOptions};
 use std::sync::Arc;
@@ -62,7 +60,7 @@ impl Device {
     }
     #[inline]
     pub fn max_buffer_length(&self) -> u64 {
-        self.inner.max_buffer_length() as u64
+        self.inner.max_buffer_length()
     }
     #[inline]
     pub fn recommended_max_working_set_size(&self) -> u64 {
